+++ packages/filesystem/src/common/filesystem.ts
@@ -134,7 +134,7 @@ export interface FileSystem extends JsonRpcServer<FileSystemClient> {
      * If the URI is not a file URI, undefined is returned.
      *
      * USE WITH CAUTION: You should always prefer URIs to paths if possible, as they are
-     * portable and platform independent. Pathes should only be used in cases you directly
+     * portable and platform independent. Paths should only be used in cases you directly
      * interact with the OS, e.g. when running a command on the shell.
      */
     getFsPath(uri: string): Promise<string | undefined>